Modelagem da base de dados e configuração das Model

+++ database/migrations/2023_04_24_153247_create_debits_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
                 ->onUpdate('cascade')
                 ->onDelete('cascade');
             $table->text('description')->nullable();
-            $table->double('value');
+            $table->decimal('value');
             $table->timestamps();
             $table->softDeletes();
         });
